+++ src/dataprotection/azext_dataprotection/tests/latest/test_dataprotection_backup_instance_operations.py
@@ -279,3 +279,52 @@ class BackupInstanceOperationsScenarioTest(ScenarioTest):
         # Once protection elsewhere is stopped, we can resume protection on the undeleted BI
         time.sleep(60)
         reset_softdelete_base_state(test)
+
+    @AllowLargeResponse()
+    def test_dataprotection_backup_instance_update_aks_configuration(test):
+        # Update with AKS backup configuration using simple az CLI commands.
+        test.kwargs.update({
+            'location': 'eastus2euap',
+            'rg': 'clitest-dpp-rg',
+            'vaultName': 'clitest-bkp-vault-aks-donotdelete',
+            'policyId': '/subscriptions/38304e13-357e-405e-9e9a-220351dcce8c/resourceGroups/clitest-dpp-rg/providers/Microsoft.DataProtection/backupVaults/clitest-bkp-vault-aks-donotdelete/backupPolicies/akspolicy',
+            'dataSourceType': 'AzureKubernetesService',
+            'aksClusterName': 'clitest-cluster1-donotdelete',
+            'aksClusterId': '/subscriptions/38304e13-357e-405e-9e9a-220351dcce8c/resourceGroups/oss-clitest-rg/providers/Microsoft.ContainerService/managedClusters/clitest-cluster1-donotdelete',
+            'friendlyName': 'clitest-cluster1-donotdelete\\clitestsabidonotdelete',
+            'backupInstanceName': 'clitest-cluster1-donotdelete-clitest-cluster1-donotdelete-bd3ae6fd-0d5f-4f76-85b4-8d7f0fe7006c'
+        })
+
+        # Fetch original BI backupDatasourceParametersList (if any) to allow resetting later
+        original_bi = test.cmd('az dataprotection backup-instance show -g "{rg}" --vault-name "{vaultName}" --name "{backupInstanceName}"').get_output_in_json()
+        originalBackupConfig = original_bi['properties']['policyInfo']['policyParameters'].get('backupDatasourceParametersList')[0]
+        test.kwargs.update({
+            'backupConfig': originalBackupConfig
+        })
+
+        # Generate the AKS backup configuration with specific values in a single command to make the change observable
+        new_backup_config_json = test.cmd('az dataprotection backup-instance initialize-backupconfig --datasource-type AzureKubernetesService '
+                                          '--included-namespaces "nsA" "nsB" --label-selectors "app=web" --excluded-resource-types "ResourceX" '
+                                          '--include-cluster-scope-resources false --snapshot-volumes false').get_output_in_json()
+        test.kwargs.update({
+            'tempBackupConfig': new_backup_config_json
+        })
+
+        # Apply temp configuration
+        test.cmd('az dataprotection backup-instance update -g "{rg}" --vault-name "{vaultName}" --backup-instance-name "{backupInstanceName}" --aks-backup-configuration "{tempBackupConfig}"', checks=[
+            test.check('name', "{backupInstanceName}")
+        ])
+
+        # Fetch the BI and verify that the backupDatasourceParametersList was updated to reflect the AKS config
+        test.cmd('az dataprotection backup-instance show -g "{rg}" --vault-name "{vaultName}" --name "{backupInstanceName}"', checks=[
+            test.check("properties.policyInfo.policyParameters.backupDatasourceParametersList[0].includedNamespaces", ['nsA', 'nsB']),
+            test.check("properties.policyInfo.policyParameters.backupDatasourceParametersList[0].labelSelectors", ['app=web']),
+            test.check("properties.policyInfo.policyParameters.backupDatasourceParametersList[0].excludedResourceTypes", ['ResourceX']),
+            test.check("properties.policyInfo.policyParameters.backupDatasourceParametersList[0].includeClusterScopeResources", False),
+            test.check("properties.policyInfo.policyParameters.backupDatasourceParametersList[0].snapshotVolumes", False)
+        ])
+        
+        # Reset to original configuration
+        test.cmd('az dataprotection backup-instance update -g "{rg}" --vault-name "{vaultName}" --backup-instance-name "{backupInstanceName}" --aks-backup-configuration "{backupConfig}"', checks=[
+            test.check('name', "{backupInstanceName}")
+        ])